restructure packages + rename SolCapServicesInfo to VCAPServicesInfo
